sql alchemy

+++ app.py
@@ -1,12 +1,16 @@
 from flask import Flask, render_template, request, redirect, url_for, flash
 from werkzeug.utils import secure_filename
-from models import db, Letter
+from models import Letter, db
 import os
 import logging
+from flask_migrate import Migrate
 
-app = Flask(__name__)
+app = Flask(__name__, static_folder='static')
 app.config.from_object('config.Config')
+
 db.init_app(app)
+
+migrate = Migrate(app, db)
 
 logging.basicConfig(level=logging.INFO)
 
@@ -29,6 +33,7 @@ def write_letter():
     if request.method == 'POST':
         name = request.form['name']
         content = request.form['content']
+        password = request.form['password']  # 비밀번호 필드 추가
         image = request.files['image']
 
         if image and image.filename and allowed_file(image.filename):
@@ -42,12 +47,12 @@ def write_letter():
             image_path = None
             logging.info("No image uploaded or invalid file type")
 
-        new_letter = Letter(name=name, content=content, image=image_path)
+        new_letter = Letter(name=name, content=content, image=image_path, password=password)
         db.session.add(new_letter)
         db.session.commit()
 
         flash('편지가 성공적으로 저장되었습니다!', 'success')
-        return redirect(url_for('letter_list'))
+        return redirect(url_for('index'))
 
     return render_template('write_letter.html')
 
@@ -55,9 +60,13 @@ def write_letter():
 def delete_letter(letter_id):
     letter = Letter.query.get(letter_id)
     if letter:
-        db.session.delete(letter)
-        db.session.commit()
-        flash('편지가 성공적으로 삭제되었습니다!', 'success')
+        password = request.form['password']  # 비밀번호 확인
+        if password == letter.password:
+            db.session.delete(letter)
+            db.session.commit()
+            flash('편지가 성공적으로 삭제되었습니다!', 'success')
+        else:
+            flash('비밀번호가 일치하지 않습니다.', 'error')
     else:
         flash('편지를 찾을 수 없습니다.', 'error')
     return redirect(url_for('letter_list'))
@@ -67,26 +76,26 @@ def edit_letter(letter_id):
     letter = Letter.query.get(letter_id)
 
     if request.method == 'POST':
-        letter.name = request.form['name']
-        letter.content = request.form['content']
+        password = request.form['password']  # 비밀번호 확인
+        if password == letter.password:
+            letter.name = request.form['name']
+            letter.content = request.form['content']
 
-        if request.files.get('image'):
-            image = request.files['image']
-            if allowed_file(image.filename):
-                filename = secure_filename(image.filename)
-                image_path = os.path.join('uploads', filename)
-                full_path = os.path.join(app.config['UPLOAD_FOLDER'], filename)
-                image.save(full_path)
-                letter.image = image_path
+            if request.files.get('image'):
+                image = request.files['image']
+                if allowed_file(image.filename):
+                    filename = secure_filename(image.filename)
+                    full_path = os.path.join(app.config['UPLOAD_FOLDER'], filename)
+                    image.save(full_path)
+                    letter.image = filename
 
-        db.session.commit()
-
-        flash('편지가 성공적으로 편집되었습니다!', 'success')
+            db.session.commit()
+            flash('편지가 성공적으로 편집되었습니다!', 'success')
+        else:
+            flash('비밀번호가 일치하지 않습니다.', 'error')
         return redirect(url_for('letter_list'))
 
     return render_template('edit_letter.html', letter=letter)
 
 if __name__ == '__main__':
-    with app.app_context():
-        db.create_all()
     app.run(debug=True)
